# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Kumquat API
  version: 1.0.0
  description: Northbound business API. Engine custom resources are internal execution projections.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      security: []
      operationId: login
      requestBody: {$ref: '#/components/requestBodies/Login'}
      responses: {'200': {$ref: '#/components/responses/Login'}, '400': {$ref: '#/components/responses/Error'}, '401': {$ref: '#/components/responses/Error'}}
  /auth/register:
    post:
      security: []
      operationId: register
      requestBody: {$ref: '#/components/requestBodies/Register'}
      responses: {'200': {$ref: '#/components/responses/User'}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /auth/me:
    get: {operationId: getCurrentUser, responses: {'200': {$ref: '#/components/responses/User'}, '401': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
  /auth/change-password:
    put:
      operationId: changePassword
      requestBody: {$ref: '#/components/requestBodies/ChangePassword'}
      responses: {'200': {$ref: '#/components/responses/Message'}, '400': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}}
...
  /modules:
    get: {operationId: listModules, responses: {'200': {$ref: '#/components/responses/ModuleList'}}}
    post:
      operationId: createModule
      requestBody: {$ref: '#/components/requestBodies/ModuleCreate'}
      responses: {'200': {$ref: '#/components/responses/Module'}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /modules/{id}:
    parameters: [{$ref: '#/components/parameters/PublicID'}]
    get: {operationId: getModule, responses: {'200': {$ref: '#/components/responses/Module'}, '404': {$ref: '#/components/responses/Error'}}}
    put:
      operationId: updateModule
      requestBody: {$ref: '#/components/requestBodies/ModuleUpdate'}
      responses: {'200': {$ref: '#/components/responses/Module'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
    delete: {operationId: deleteModule, responses: {'200': {$ref: '#/components/responses/Message'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    PublicID: {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/PublicID'}}
...
  requestBodies:
    Login: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LoginRequest'}}}}
    Register: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RegisterRequest'}}}}
    ChangePassword: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ChangePasswordRequest'}}}}
...
    ModuleCreate: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ModuleCreate'}}}}
    ModuleUpdate: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ModuleUpdate'}}}}
...
  schemas:
    PublicID: {type: string, minLength: 3, maxLength: 40, pattern: '^[a-z]+_[a-f0-9]+$'}
    LoginRequest: {type: object, additionalProperties: false, required: [username, password], properties: {username: {type: string}, password: {type: string, format: password}}}
    RegisterRequest: {type: object, additionalProperties: false, required: [username, email, password], properties: {username: {type: string, minLength: 3, maxLength: 32}, email: {type: string, format: email}, password: {type: string, format: password, minLength: 6, maxLength: 32}, nickname: {type: string}}}
    ChangePasswordRequest: {type: object, additionalProperties: false, required: [oldPassword, newPassword], properties: {oldPassword: {type: string, format: password}, newPassword: {type: string, format: password, minLength: 6, maxLength: 32}}}
...
    ModuleCreate: {type: object, additionalProperties: false, required: [name], properties: {name: {type: string}, parentId: {$ref: '#/components/schemas/PublicID'}, sort: {type: integer}}}
    ModuleUpdate: {type: object, additionalProperties: false, properties: {name: {type: string}, parentId: {$ref: '#/components/schemas/PublicID'}, sort: {type: integer}}}
...
    ErrorEnvelope: {type: object, additionalProperties: false, required: [code, message], properties: {code: {type: integer}, message: {type: string}}}
    RoleDTO: {type: object, additionalProperties: false, required: [id, name, created_at, updated_at], properties: {id: {type: integer, format: int64}, name: {type: string}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}}
    ModuleDTO: {type: object, additionalProperties: false, required: [id, name, level, sort, path, created_at, updated_at], properties: {id: {$ref: '#/components/schemas/PublicID'}, name: {type: string}, parent_id: {$ref: '#/components/schemas/PublicID'}, level: {type: integer}, sort: {type: integer}, path: {type: string}, children: {type: array, items: {$ref: '#/components/schemas/ModuleDTO'}}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}}
    UserDTO: {type: object, additionalProperties: false, required: [id, username, email, nickname, role_id, created_at, updated_at], properties: {id: {type: integer, format: int64}, username: {type: string}, email: {type: string, format: email}, nickname: {type: string}, role_id: {type: integer, format: int64}, role: {$ref: '#/components/schemas/RoleDTO'}, module_id: {type: integer, format: int64, nullable: true}, module: {$ref: '#/components/schemas/ModuleDTO'}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}}
...
    MessageEnvelope: {type: object, additionalProperties: false, required: [code, message], properties: {code: {type: integer, enum: [0]}, message: {type: string}}}
    LoginEnvelope: {type: object, additionalProperties: false, required: [code, message, data], properties: {code: {type: integer, enum: [0]}, message: {type: string}, data: {type: object, additionalProperties: false, required: [token, user], properties: {token: {type: string}, user: {$ref: '#/components/schemas/UserDTO'}}}}}
    UserEnvelope: {type: object, additionalProperties: false, required: [code, message, data], properties: {code: {type: integer, enum: [0]}, message: {type: string}, data: {$ref: '#/components/schemas/UserDTO'}}}
...
    ModuleEnvelope: {type: object, additionalProperties: false, required: [code, message, data], properties: {code: {type: integer, enum: [0]}, message: {type: string}, data: {$ref: '#/components/schemas/ModuleDTO'}}}
...
    ModuleListEnvelope: {type: object, additionalProperties: false, required: [code, message, data], properties: {code: {type: integer, enum: [0]}, message: {type: string}, data: {type: array, items: {$ref: '#/components/schemas/ModuleDTO'}}}}
...
  responses:
    Login: {description: Login, content: {application/json: {schema: {$ref: '#/components/schemas/LoginEnvelope'}}}}
    User: {description: User, content: {application/json: {schema: {$ref: '#/components/schemas/UserEnvelope'}}}}
...
    Module: {description: Module, content: {application/json: {schema: {$ref: '#/components/schemas/ModuleEnvelope'}}}}
    ModuleList: {description: Modules, content: {application/json: {schema: {$ref: '#/components/schemas/ModuleListEnvelope'}}}}
...
    Message: {description: Message, content: {application/json: {schema: {$ref: '#/components/schemas/MessageEnvelope'}}}}
...
    Error: {description: Error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}